backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
openai==1.12.0
beautifulsoup4==4.12.3
lxml==5.1.0
pypdf==4.0.1
aiofiles==23.2.1
